Former-commit-id: 228b7951175d524372bd3c732dec9bfd2beea409
Former-commit-id: 086cea07500f6f841f7f10623e54529a46fd1f8c

+++ core/domain/interface/member/member_repo.go
@@ -72,7 +72,7 @@ type IMemberRepo interface {
 	GetMemberIdByUser(user string) int64
 
 	// 根据编码获取会员
-	GetMemberIdByCode(code string) int
+	GetMemberIdByCode(code string) int64
 
 	// 用户名是否存在
 	CheckUsrExist(user string, memberId int64) bool
